refactor(client): extract xclient call methods into xclient_call.go
Move core invocation (Go, Call, Oneshot, SendRaw) and their helpers
(setServerTimeout, uncoverError, contextCanceled, wrapCall, wrapSendRaw)
out of xclient.go into xclient_call.go. Same package, mechanical move.
